+++ api/models/Comment.go
@@ -109,7 +109,7 @@ func (c *Comment) DeleteAComment(db *gorm.DB) (int64, error) {
 }
 
 //When a user is deleted, we also delete the comments that the user had
-func (c *Comment) DeleteUserComments(db *gorm.DB, uid uint64) (int64, error) {
+func (c *Comment) DeleteUserComments(db *gorm.DB, uid uint32) (int64, error) {
 	comments := []Comment{}
 	db = db.Debug().Model(&Comment{}).Where("user_id = ?", uid).Find(&comments).Delete(&comments)
 	if db.Error != nil {
